+++ tests/test_webhook_validation.py
@@ -6,7 +6,7 @@ import pytest
 from fastapi import HTTPException, WebSocketDisconnect
 from twilio.request_validator import RequestValidator
 
-from tac.server.webhook import (
+from tac.server.signature_validation import (
     _build_url,
     _build_websocket_url_and_params,
     _http_scheme_to_ws,
